Add JWT ExtendedL

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from sqlalchemy import exc
 from models.item import ItemModel
 
@@ -9,13 +9,14 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help="This Field cannot be left blank!")
     parser.add_argument('store_id', type=int, required=True, help="Every item needs an store id")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json(), 200
         return {'message': 'Item not found'}, 400
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': f"An item with name {name} already exists."}, 400
@@ -30,7 +31,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 400
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -56,5 +62,11 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}
+        else:
+            return {'items': [item['name'] for item in items], 'message': 'More data available if you log in'}, 200
